requirements.txt
pytest==3.6.1
pytest-xdist==1.22.0
h5py==2.7.1
numpy==1.14.0
tensorflow==1.8.0
keras==2.1.2
progressbar33==2.4.0

